Refactor
This was the start of a major refactor. There is now a Deck object which
creates 52 instances of the new Card object to form a deck. It has a
shuffle and Draw function. There is a single Hand object for Dealer ,
Player, and Community that handles what cards are in the hand. Cards
Object has A suit and face value , a toString method , and a
GetSpriteIndex function which determines the sprite via suit/face and if
face up or face down