Refactor
Removal of lots of code , Can now Click deal and cards are dealt , then
click the bet button to skip through the game states and see the hands
and winner in debug text. Can Also drage chips to ANTE and Bonus spots
and chip totalis calculated. Player Bank is also show but not yet added
to or subtracted. Other Bets and Payouts not refactored yet