Chore/refactor (#10)
* wip

* Refactor Of Poker Scripts

Created a couple Poker Objects to contain the Poker related functions ,
still needs additional refactor including moving card related functions
to a card object

* Further refactor

Started a Deck Object and initial refactor into a class like object

* More refactoring

Bug introduced where it seems like 6 cards are being highlighted

* Refactor

This was the start of a major refactor. There is now a Deck object which
creates 52 instances of the new Card object to form a deck. It has a
shuffle and Draw function. There is a single Hand object for Dealer ,
Player, and Community that handles what cards are in the hand. Cards
Object has A suit and face value , a toString method , and a
GetSpriteIndex function which determines the sprite via suit/face and if
face up or face down

* Refactor

Wired up Refactored code with a Poker and Poker Evaluator class , added
debug winner text , added game_controller and moved logic from btn_bet
to it

* Refactor

Removal of lots of code , Can now Click deal and cards are dealt , then
click the bet button to skip through the game states and see the hands
and winner in debug text. Can Also drage chips to ANTE and Bonus spots
and chip totalis calculated. Player Bank is also show but not yet added
to or subtracted. Other Bets and Payouts not refactored yet

* Small cleanup

* small rename

* Refactor

Fold and Check buttons work , no validation currently implemented

* Payouts Refactored

No payout visual indicator or animation but user total is updated with
winnings , fixed bug where the bets were being set before they were
calculated leading to a one behind scenario

* Refactor

Added Payout Animations back in , issue with payout animation for 1000
to 1 the animation glitches out. Will need to adjust animation maybe
switch to a lerp

* Rename Controller

---------

Co-authored-by: James Preston <dxsolo@gmail.com>